split service into rate and email parts
